Try to add mapped typogarphic glyph to sequence when using "Add" button on Character Map
