+++ CharacterMap/CharacterMap/CharacterMap.csproj
@@ -341,6 +341,7 @@
     <Compile Include="ViewModels\FontMapViewModel.cs" />
     <Compile Include="Models\GlyphAnnotation.cs" />
     <Compile Include="Models\PrintLayout.cs" />
+    <Compile Include="ViewModels\FontMapViewModel.SubModels.cs" />
     <Compile Include="ViewModels\GlyphFileNameViewModel.cs" />
     <Compile Include="ViewModels\PrintViewModel.cs" />
     <Compile Include="ViewModels\QuickCompareViewModel.cs" />
